Validate the to option shape at the declaration
